Move erf and erf_inv out of microfacet. They could be useful in other contexts.
